debug LTR; update early stop

+++ openpto/method/Models/LTR.py
@@ -29,7 +29,7 @@ class pointwiseLTR(optModel):
         super().__init__(optSolver, processes, solve_ratio)
         # solution pool
         n_vars = optSolver.num_vars
-        self.solpool = np.empty((0, n_vars), dtype=np.float)
+        self.solpool = np.empty((0, n_vars), dtype=np.float32)
 
     def forward(self, problem, coeff_hat, coeff_true, params, **hyperparams):
         """
@@ -104,7 +104,7 @@ class pairwiseLTR(optModel):
         super().__init__(optSolver, processes, solve_ratio)
         # solution pool
         n_vars = optSolver.num_vars
-        self.solpool = np.empty((0, n_vars), dtype=np.float)
+        self.solpool = np.empty((0, n_vars), dtype=np.float32)
 
     def forward(self, problem, coeff_hat, coeff_true, params, **hyperparams):
         """
@@ -208,7 +208,7 @@ class listwiseLTR(optModel):
         self.tau = tau
         # solution pool
         n_vars = optSolver.num_vars
-        self.solpool = np.empty((0, n_vars), dtype=np.float)
+        self.solpool = np.empty((0, n_vars), dtype=np.float32)
 
     def forward(self, problem, coeff_hat, coeff_true, params, **hyperparams):
         """
